简单适配600px以下设备

+++ src/pages/DraftPreview.jsx
@@ -4,43 +4,40 @@ import { useState, useEffect, Fragment, useContext } from 'react'
 import { getDraft, deletePost } from '../axios/request'
 import moment from 'moment'
 import { AuthContext } from '../context/authContext.js'
-import { UserOutlined, EditTwoTone, DeleteTwoTone } from '@ant-design/icons';
-import { Avatar, Modal, message, PageHeader, Tooltip } from 'antd';
+import { UserOutlined, EditTwoTone, DeleteTwoTone, ExclamationCircleOutlined } from '@ant-design/icons';
+import { Avatar, Modal, message, PageHeader, Tooltip, Image } from 'antd';
 import { loginErrorFn } from '../units/errorFn'
 import { scrollToTop } from '../units/scrollToTop'
 import Footer from '../components/Footer'
 
 export default function Single() {
-  useEffect(() => {
-    scrollToTop()
-  })
 
   const [post, setPost] = useState({})
-  const [open, setOpen] = useState(false)
-  const [confirmLoading, setConfirmLoading] = useState(false)
+  const [visible, setVisible] = useState(false)
+  const [previewSrc, setPreviewSrc] = useState('error')
   const navigate = useNavigate()
   let draftId = useParams().id
   const { currentUsername, logout } = useContext(AuthContext)
-  // 确认框的异步确认回调
-  const handleOk = async () => {
-    setConfirmLoading(true);
-    try {
-      const res = await deletePost(draftId)
-      if (res.status === 200) {
-        message.success('删除成功 !')
-        navigate(-1, { replace: true })
-      }
-    } catch (error) {
-      loginErrorFn(error, Modal, message, navigate, logout)
-    }
-    setOpen(false)
-    setConfirmLoading(false)
-  };
-  const handleCancel = () => {
-    setOpen(false);
-  };
+  // 确认框的确认回调
   const onDelete = async function () {
-    setOpen(true)
+    Modal.confirm({
+      title: 'Tips',
+      icon: <ExclamationCircleOutlined />,
+      content: '是否确认删除, 无法恢复 !',
+      okText: '确认',
+      cancelText: '取消',
+      onOk: async () => {
+        try {
+          const res = await deletePost(draftId)
+          if (res.status === 200) {
+            message.success('删除成功 !')
+            navigate(-1, { replace: true })
+          }
+        } catch (error) {
+          loginErrorFn(error, Modal, message, navigate, logout)
+        }
+      },
+    })
   }
   useEffect(() => {
     const getEffect = async () => {
@@ -56,21 +53,36 @@ export default function Single() {
       }
     }
     getEffect()
+    scrollToTop()
     // eslint-disable-next-line
   }, [draftId])
 
+  const onImgPreview = function (e) {
+    if (e.target.nodeName === 'IMG') {
+      setVisible(true)
+      setPreviewSrc(e.target.src)
+    }
+  }
+
   return (
     <Fragment>
-      {/* 确认弹窗 */}
-      <Modal
-        title="Tips"
-        open={open}
-        onOk={handleOk}
-        confirmLoading={confirmLoading}
-        onCancel={handleCancel}
-      >
-        <p>{'是否确认删除该文章 ( 无法恢复 ) !'}</p>
-      </Modal>
+      {/* 预览图，默认隐藏 */}
+      <Image
+        width={200}
+        style={{
+          display: 'none',
+        }}
+        src={previewSrc}
+        preview={{
+          visible,
+          scaleStep: 0.5,
+          src: previewSrc,
+          onVisibleChange: (value) => {
+            setVisible(value);
+          },
+        }}
+      />
+
       <div className='draft-preview'>
         <div className='header'>
           <PageHeader
@@ -82,7 +94,7 @@ export default function Single() {
         {/* 主题内容 */}
         <div className='content'>
           {/* 异步加载post，因此更新后再渲染数据 */}
-          {post?.img && <img src={post.img} alt={post?.title} className='opacity img' onLoad={(e) => e.target.className = 'unopacity img'} />}
+          {post?.img && <img src={post.img} alt={post?.title} className='opacity img' onLoad={(e) => e.target.className = 'unopacity img'} onClick={onImgPreview}/>}
           {/* 头像 */}
           <div className='user'>
             {post.userImg === 'null' ? <Avatar size={50} icon={<UserOutlined />} /> : <img src={post?.userImg} alt="" />}
@@ -106,7 +118,7 @@ export default function Single() {
             }
           </div>
           <h1>{post?.title}</h1>
-          <p dangerouslySetInnerHTML={{ __html: post?.description }}>
+          <p className='desc' onClick={onImgPreview} dangerouslySetInnerHTML={{ __html: post?.description }}>
           </p>
         </div>
       </div>
